requirements.txt
aiohttp>=3.9.0,<4.0.0
feedparser>=6.0.0,<7.0.0
python-dateutil>=2.8.0,<3.0.0
httpx>=0.27.0,<1.0.0
pydantic>=2.0.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
